+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features / enhancements
 
 * [booking] add calendar default view option in booking config
+* [booking] allow to delete bookingAuthorisations
 
 ### Fixes